Hard-code CakePHP 4 extension as php
I think it could be problematic changing the default of the CakePHP 3 config to PHP - it might cause the plugin view navigating / completion to not work on CakePHP 3.

So, instead of doing that, let's just split out the extension config and make it unconfigurable for Cake 4.

This also makes the old CakePHP 3 view tests pass, which makes me feel better about releasing with this.