Split up controller completion contributor
The controller completion contributor seems to be failing due to a timing
issue due to timing out during load all the models and components in a
real CakePHP 2 project in newer versions of PhpStorm.

Splitting up the model and component completion methods into two separate
contributors doesn't hit that issue.