Limit view helper completion
The new ViewHelperCompletionContributor is triggering in too many places.

Limit it to only trigger in view helpers. The method used for this is
to check the sizes of the list of all view helpers. This isn't the most efficient
method - should probably use a heuristic to avoid having to load the
list unless it's necessary.