Prefix namespace with backslash
The appNamespace var wasn't prefixed with backslash, but that's required in a number of places where it's used.

We already add the backslash on save, but in case someone has saved a namespace without a backslash, change the getter method for it on the Settings to prefix the value with a backslash before returning it.