Rename templatePath -> pluginPath
we want the ability to get the assets underneath the plugins as well - which is configued by plugins/MyPlugin/webroot/css/my_css.css, so it's not just templates that we'll need to configure.

That does mean we'll need to expand the path based on the CakepHP version to get the templates dir (So plugins/MyPlugin/src/Template for Cake 3, plugins/MyPlugin/templates for Cake 4+), but se la vi.