Use PluginConfig instead of PluginEntries
PluginConfig is immutable, and it supersedes the PluginEntry. Reserve the PluginEntry as a mutable class for editing, and use the PluginConfig everywehre else.

Also, make it so we don't always combine the legacy pluginNamespaces with the new plugin configs stored in the settings file, so that in the usual case where it hasn't been set (or if the config is updated to convert it) that we don't combine the lists always.