Change the default shortcut for toggling between controller/view
On Linux, the default shortcut key does not work.

Should probably use the RelatedItem interface instead, which can
use the built-in "ctrl-alt-Home" / "shift-meta-up" shortcuts that
go to related items.

But, that would require some major surgery. So not doing that
at the moment.

So for now, change the shortcut key to match that related item
pattern, but add the Shift key by default.