Add Settings.cake3AutoDetect property
In the settings form, we need to read / write the raw value of the
cake3 autodetect value, which was formerly named cake3Enabled.

Since, we're using the cake3Enabled property in the Settings form now,
that leads to a problem with settings the autodetected value to false
when configuring Settings for new projects.

So, expose a new getter for the cake3AutoDetect, and set the state.cake3Enabled
value from it.